Fix Prometheus register view error
Registering multiple views with the Prometheus exporter resulted
in the `descriptor %s already exists with the same fully-qualified
name and const label values` error from the Prometheus client.

The Prometheus client code defines the identity of the collector
by the sum of the description IDs in the collector, so changing
the set of descriptions in a collector to add a view changes the
collector's ID, so unregistering the collector didn't do anything,
so re-registering the collector caused the registry to complain
about duplicate descriptions.

This change moves the unregistering of the collector to before
adding the new views so that everything works more smoothly.